group_task_1/pyproject.toml
[project]
name = "group-task-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "librosa>=0.11.0",
    "numpy>=2.0.0",
    "pandas>=2.0.0",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "pytorch-lightning>=2.2.0",
    "wandb>=0.16.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
    "soundfile>=0.12.1",
    "editdistance>=0.8.1",
]
